scrape/requirements-api.txt
# Web Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0.23
alembic>=1.12.1
asyncpg>=0.29.0
# psycopg2-binary>=2.9.9  # PostgreSQL 설치 후 별도 설치

# Data Processing
pandas>=2.1.3
numpy>=1.25.2

# Validation & Serialization
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP Client
httpx>=0.25.2
aiofiles>=23.2.1

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0

# Development Tools
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.1

# Utilities
python-dotenv>=1.0.0
click>=8.1.7
rich>=13.7.0
tqdm>=4.66.1

# Korean Text Processing (Python 3.12 호환 버전으로 변경)
# konlpy==0.6.0  # Python 3.12 호환성 문제로 주석 처리
# hanja==0.13.0  # Python 3.12 호환성 문제로 주석 처리

# Existing crawler dependencies
requests
beautifulsoup4
lxml
aiohttp
psutil